Saving before taking different approach
Python is doing too many allocations, our simple allocator is not good
enough in terms of performance, so let's try to hijack python's one.

The idea I have right now is to take the arena allocator that they
define, and add the layer of indirection on top of it rather than below
it.